fix(eslint-errors): cookie-consent
